requirements.txt
Django==4.0.2
Pillow==9.0.1
django-tables2==2.4.1
django-crispy-forms==1.14.0
psycopg2-binary==2.9.3
django-storages==1.13.1
boto3==1.24.71
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-filter==22.1
requests==2.28.1
beautifulsoup4==4.11.1